+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- rename option from `exclude` to `exclude-by-name`
 
 ## [3.1.1]- 2018-12-09
 ### Fixed